requirements.txt
lxml
flair
scispacy==0.2.4
tqdm
